Adds backward compatibility to the BootManager for older bootcds.  Rather than failing to load the new pciscan library, it will succeed with the old behavior.  In particular some devices will fail to be detected using the old bootcd.  But, this shouldn't be that much of a problem.
Daniel has proposed that the pypciscan.so library simply be re-written in
python by parsing /proc/bus/pci/devices to create a structure like:

    pypciscan.get_devices() 
		returns {'<domain>:<bus>:<dev>.<func>' : 
					(<vendor id>, <device id>, <subvendor id>, <subdevice id>, <device class << 8 | progif>)} 
		for all PCI devices.

I agree that this would be a better approach.  Would have to check that the
older bootcd kernel exports all of this information, then rewrite the module.